# api/openapi.yaml
openapi: "3.1.0"
info:
  title: "Kinetic v3 API"
  version: "0.1.0"
  description: |
    REST API for the Kinetic v3 Rocket League esports league management platform.

    ## Authentication
    Most read endpoints are publicly accessible. Write and sensitive operations
    require authentication via either a session cookie (set by the OAuth login flow)
    or an HTTP Bearer token (API token issued via `/v1/api-tokens`).

    ## Permission model
    Protected routes enforce a resource/action pair checked against the caller's
    roles. Available resources are defined in `authz.resources` and include
    `league`, `franchise`, `club`, `team`, `player`, `roster_membership`,
    `role_assignment`, `queue`, `queue_entry`, `queue_ban`, `scrim`,
    `scrim_promotion`, `result_submission`, `result_override`, `replay_evidence`,
    `platform_account_link`, `player_rating`, `rating_adjustment`,
    `matchmaking_decision`, `exception_ticket`, `season`, `schedule_group`,
    `fixture`, `match`, `eligibility`, `player_stat`, `skill_group`,
    `skill_group_transition`, `organization_config`, `api_token`, and
    `player_notification`.
servers:
  - url: "http://localhost:8080"
    description: "Local development server"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kinetic_session
      description: "Signed session token set by the OAuth login flow."
    bearerToken:
      type: http
      scheme: bearer
      description: "Long-lived API token issued via POST /v1/api-tokens."
  schemas:
...
    QueueEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        isActive:
          type: boolean
        stage:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        stageAt:
          type: string
          format: date-time
        leftAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - queueId
        - teamId
        - isActive
        - stage
        - createdAt
        - stageAt
    QueueBan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
        bannedByActor:
          type: string
        banReason:
          type: string
        isActive:
          type: boolean
        bannedAt:
          type: string
          format: date-time
        unbannedByActor:
          type: string
          nullable: true
        unbanReason:
          type: string
          nullable: true
        unbannedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - queueId
        - playerId
        - bannedByActor
        - banReason
        - isActive
        - bannedAt

    # ── Scrim ──────────────────────────────────────────────────────────────────
    Scrim:
      type: object
      properties:
        id:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
        homeTeamId:
          type: integer
          format: int64
        awayTeamId:
          type: integer
          format: int64
        state:
          type: string
          description: "State machine value (e.g. popped, checked_in, in_progress, complete, cancelled)."
        lobbyName:
          type: string
          nullable: true
        lobbyPassword:
          type: string
          nullable: true
        poppedAt:
          type: string
          format: date-time
          nullable: true
        homeCheckedInAt:
          type: string
          format: date-time
          nullable: true
        awayCheckedInAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - queueId
        - homeTeamId
        - awayTeamId
        - state
        - createdAt
...
    ProcessQueuePromotionsResult:
      type: object
      properties:
        processedQueues:
          type: integer
          format: int32
        promotionsCreated:
          type: integer
          format: int32
        conflicts:
          type: integer
          format: int32
        poppedScrimIds:
          type: array
          items:
            type: integer
            format: int64
      required:
        - processedQueues
        - promotionsCreated
        - conflicts
        - poppedScrimIds
    PromotionProcessingRun:
      type: object
      properties:
        id:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
          nullable: true
        processedQueues:
          type: integer
          format: int32
        promotionsCreated:
          type: integer
          format: int32
        conflicts:
          type: integer
          format: int32
        durationMs:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - processedQueues
        - promotionsCreated
        - conflicts
        - durationMs
        - createdAt

    # ── Ratings ────────────────────────────────────────────────────────────────
    PlayerRating:
      type: object
      properties:
        id:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
        contextKey:
          type: string
          description: "Opaque key identifying the rating context (e.g. queue slug)."
        rating:
          type: integer
          format: int32
        uncertainty:
          type: integer
          format: int32
        matchesPlayed:
          type: integer
          format: int32
        lastCompetedAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - playerId
        - contextKey
        - rating
        - uncertainty
        - matchesPlayed
        - isActive
        - updatedAt
    RatingAdjustment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        actorPlayerId:
          type: integer
          format: int64
        targetPlayerId:
          type: integer
          format: int64
        contextKey:
          type: string
        previousRating:
          type: integer
          format: int32
        newRating:
          type: integer
          format: int32
        previousUncertainty:
          type: integer
          format: int32
        newUncertainty:
          type: integer
          format: int32
        previousMatchesPlayed:
          type: integer
          format: int32
        newMatchesPlayed:
          type: integer
          format: int32
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - actorPlayerId
        - targetPlayerId
        - contextKey
        - previousRating
        - newRating
        - previousUncertainty
        - newUncertainty
        - previousMatchesPlayed
        - newMatchesPlayed
        - reason
        - createdAt
    MatchmakingDecision:
      type: object
      properties:
        id:
          type: integer
          format: int64
        scrimId:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
        queueWaitSeconds:
          type: integer
          format: int32
        waitSkewSeconds:
          type: integer
          format: int32
        expansionStage:
          type: integer
          format: int32
        ratingSpread:
          type: integer
          format: int32
        homeTeamRating:
          type: integer
          format: int32
        awayTeamRating:
          type: integer
          format: int32
        crossGroup:
          type: boolean
        orderingStrategy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - scrimId
        - queueId
        - queueWaitSeconds
        - waitSkewSeconds
        - expansionStage
        - ratingSpread
        - homeTeamRating
        - awayTeamRating
        - crossGroup
        - orderingStrategy
        - createdAt

    # ── Results ────────────────────────────────────────────────────────────────
...
paths:
...
  /v1/queue-entries:
    get:
      operationId: listQueueEntries
      summary: "List active queue entries"
      tags:
        - queue
      responses:
        "200":
          description: "Active queue entries."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueEntry"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: enqueueTeam
      summary: "Add a team to a queue"
      tags:
        - queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
                teamId:
                  type: integer
                  format: int64
              required:
                - queueId
                - teamId
      responses:
        "201":
          description: "Queue entry created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "400":
          description: "Invalid input."
        "409":
          description: "Team already in queue."
        "503":
          description: "Hierarchy store unavailable."
    delete:
      operationId: leaveQueue
      summary: "Remove a team from a queue"
      tags:
        - queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
                teamId:
                  type: integer
                  format: int64
              required:
                - queueId
                - teamId
      responses:
        "200":
          description: "Updated queue entry."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "400":
          description: "Invalid input."
        "503":
          description: "Hierarchy store unavailable."
    patch:
      operationId: advanceQueueEntryStage
      summary: "Advance the matchmaking expansion stage for a queue entry"
      tags:
        - queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
                teamId:
                  type: integer
                  format: int64
                stage:
                  type: integer
                  format: int32
              required:
                - queueId
                - teamId
                - stage
      responses:
        "200":
          description: "Updated queue entry."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "400":
          description: "Invalid input."
        "503":
          description: "Hierarchy store unavailable."
  /v1/queue-bans:
    get:
      operationId: listQueueBans
      summary: "List all queue bans"
      tags:
        - queue
      responses:
        "200":
          description: "All queue bans."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueBan"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: banPlayerFromQueue
      summary: "Ban a player from a queue"
      tags:
        - queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
                playerId:
                  type: integer
                  format: int64
                actor:
                  type: string
                reason:
                  type: string
              required:
                - queueId
                - playerId
                - actor
                - reason
      responses:
        "201":
          description: "Queue ban created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueBan"
        "400":
          description: "Invalid input."
        "409":
          description: "Player already banned."
        "503":
          description: "Hierarchy store unavailable."
  /v1/queue-bans/lift:
    post:
      operationId: liftQueueBan
      summary: "Lift (unban) a player from a queue"
      tags:
        - queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
                playerId:
                  type: integer
                  format: int64
                actor:
                  type: string
                reason:
                  type: string
              required:
                - queueId
                - playerId
                - actor
                - reason
      responses:
        "200":
          description: "Updated queue ban."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueBan"
        "400":
          description: "Invalid input."
        "503":
          description: "Hierarchy store unavailable."

  # ── Scrims ─────────────────────────────────────────────────────────────────
  /v1/scrims:
    get:
      operationId: listScrims
      summary: "List all scrims"
      tags:
        - scrims
      responses:
        "200":
          description: "All scrims."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scrim"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: createScrim
      summary: "Create a scrim manually"
      tags:
        - scrims
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
                homeTeamId:
                  type: integer
                  format: int64
                awayTeamId:
                  type: integer
                  format: int64
                state:
                  type: string
              required:
                - queueId
                - homeTeamId
                - awayTeamId
                - state
      responses:
        "201":
          description: "Created scrim."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
        "400":
          description: "Invalid input."
        "409":
          description: "Conflict."
        "503":
          description: "Hierarchy store unavailable."
    patch:
      operationId: updateScrimState
      summary: "Update the state of a scrim"
      tags:
        - scrims
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scrimId:
                  type: integer
                  format: int64
                state:
                  type: string
              required:
                - scrimId
                - state
      responses:
        "200":
          description: "Updated scrim."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
        "400":
          description: "Invalid input or invalid state transition."
        "503":
          description: "Hierarchy store unavailable."
...
  /v1/scrim-promotions:
    post:
      operationId: promoteQueueToScrim
      summary: "Pop a single queue match into a scrim (requires scrim_promotion:create)"
      tags:
        - scrims
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
              required:
                - queueId
      responses:
        "201":
          description: "Created scrim."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
        "400":
          description: "Invalid input."
        "401":
          description: "Not authenticated."
        "403":
          description: "Insufficient permissions."
        "409":
          description: "No eligible pairs in queue."
        "503":
          description: "Hierarchy store unavailable."
  /v1/scrim-promotions/process:
    post:
      operationId: processQueuePromotions
      summary: "Run the full matchmaking promotion pass for a queue"
      tags:
        - scrims
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                queueId:
                  type: integer
                  format: int64
              required:
                - queueId
      responses:
        "200":
          description: "Promotion processing result."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessQueuePromotionsResult"
        "400":
          description: "Invalid input."
        "401":
          description: "Not authenticated."
        "403":
          description: "Insufficient permissions."
        "503":
          description: "Hierarchy store unavailable."
  /v1/promotion-processing-runs:
    get:
      operationId: listPromotionProcessingRuns
      summary: "List historical promotion processing runs"
      tags:
        - scrims
      responses:
        "200":
          description: "Promotion processing runs."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromotionProcessingRun"
        "503":
          description: "Hierarchy store unavailable."

  # ── Ratings ────────────────────────────────────────────────────────────────
  /v1/player-ratings:
    get:
      operationId: listPlayerRatings
      summary: "List all player ratings"
      tags:
        - ratings
      responses:
        "200":
          description: "All player ratings."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerRating"
        "503":
          description: "Hierarchy store unavailable."
  /v1/player-ratings/adjust:
    post:
      operationId: adjustPlayerRating
      summary: "Manually adjust a player's rating"
      tags:
        - ratings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actorPlayerId:
                  type: integer
                  format: int64
                targetPlayerId:
                  type: integer
                  format: int64
                contextKey:
                  type: string
                rating:
                  type: integer
                  format: int32
                uncertainty:
                  type: integer
                  format: int32
                matchesPlayed:
                  type: integer
                  format: int32
                reason:
                  type: string
              required:
                - actorPlayerId
                - targetPlayerId
                - contextKey
                - rating
                - uncertainty
                - matchesPlayed
                - reason
      responses:
        "200":
          description: "Updated player rating."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerRating"
        "400":
          description: "Invalid input."
        "503":
          description: "Hierarchy store unavailable."
  /v1/rating-adjustments:
    get:
      operationId: listRatingAdjustments
      summary: "List all rating adjustment records"
      tags:
        - ratings
      responses:
        "200":
          description: "All rating adjustments."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RatingAdjustment"
        "503":
          description: "Hierarchy store unavailable."
  /v1/matchmaking-decisions:
    get:
      operationId: listMatchmakingDecisions
      summary: "List matchmaking decision records"
      tags:
        - ratings
      responses:
        "200":
          description: "All matchmaking decisions."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchmakingDecision"
        "503":
          description: "Hierarchy store unavailable."

  # ── Results ────────────────────────────────────────────────────────────────
...